# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Crypto Payment Gateway API
  version: "1.1.0"
  description: >
    USDT payment gateway supporting BEP20 (BNB Smart Chain) and TRC20 (Tron).
    Merchant endpoints authenticate with API key + HMAC signature; dashboard
    endpoints authenticate with JWT. The target chain is chosen per-payment via
    the optional `network` field (defaults to BEP20 for backward compatibility);
    TRC20 must be enabled server-side. Amounts are always human USDT strings —
    the gateway handles each chain's decimals (BEP20=18, TRC20=6) internally.
servers:
  - url: https://gateway.example.com/api/v1
  - url: http://localhost:4000/api/v1
...
components:
  securitySchemes:
    # Two API key modes coexist. Which one applies is decided by the key itself
    # (api_keys.auth_mode), not by the caller — presenting an HMAC key without a
    # signature, or a bearer token with one, is rejected rather than downgraded.
    #
    # 1. HMAC-signed: ApiKey carries the PUBLIC id (pk_live_…), and Signature +
    #    Timestamp must accompany it. The secret never crosses the wire.
    #
    #    X-Timestamp is unix SECONDS. The signed string is the timestamp, a
    #    literal ".", then the exact raw request body:
    #        X-Signature = hex( HMAC_SHA256( apiSecret, "${X-Timestamp}.${rawBody}" ) )
    #    For a request with no body (any GET) the raw body is "", so the signed
    #    string is "${X-Timestamp}.". Serialize the body once and send those
    #    exact bytes — re-serializing changes them and invalidates the
    #    signature. Requests more than 300 s from server time are rejected as
    #    replays. Verified against backend/src/middleware/auth.ts.
    ApiKey:      { type: apiKey, in: header, name: X-Api-Key }
    Signature:   { type: apiKey, in: header, name: X-Signature }
    Timestamp:   { type: apiKey, in: header, name: X-Timestamp }
    # 2. Bearer: the X-Api-Key value IS the secret (ak_live_…), sent alone.
    #    Simpler to integrate, but the credential is on the wire every call, so
    #    such keys are never granted the payouts:write scope.
    ApiKeyBearer: { type: apiKey, in: header, name: X-Api-Key }
    BearerAuth:  { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
...
    Payment:
      type: object
      properties:
        paymentId: { type: string, example: pay_01HZY... }
        orderId:   { type: string, example: order_789 }
        amount:    { type: string, example: "50.00" }
        amountReceived: { type: string, example: "0" }
        currency:  { type: string, example: USDT }
        network:   { type: string, enum: [BEP20, TRC20, ERC20, BTC], example: BEP20 }
        asset:
          type: string
          description: >
            Ledger asset; always equal to `currency`. Both are returned because
            `currency` is the historical field name and `asset` is what the
            multi-asset endpoints key on.
          example: USDT
        address:   { type: string, description: "BEP20 0x… or TRC20 T… deposit address", example: "0xabc..." }
        qrCode:    { type: string, description: base64 data URI }
        status:    { type: string, enum: [waiting,confirming,confirmed,partial,failed,expired,swept] }
        confirmations: { type: integer }
        txHash:    { type: string, nullable: true }
        expiresAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        fiat:
          $ref: '#/components/schemas/FiatQuote'
    FiatQuote:
      type: object
      description: >
        The frozen fiat quote, present ONLY on a payment that was priced in fiat.
        Its absence means the payment was priced directly in crypto.

        The rate is captured when the payment is created and is never recomputed:
        a customer quoted 52.581765 USDT for ₹5,000 owes that amount whenever they
        pay, and it is what the QR encodes and what confirmation is measured
        against. Treat it as an audit record of a decision, not a live price.
      properties:
        currency: { type: string, example: INR, description: ISO-4217 code the merchant priced in }
        amount:   { type: string, example: "5000", description: What the merchant asked for, in that currency }
        rate:
          type: string
          example: "95.09"
          description: Price of one unit of `asset` in `currency`, at `lockedAt`.
        source:
          type: string
          nullable: true
          example: coingecko
          description: >
            Provenance, including whether the rate was served stale
            (`coingecko:stale`). A stale quote is legitimate — the gateway
            prefers a slightly old price to a failed payment — but it is never
            silent.
        lockedAt:
          type: string
          format: date-time
          nullable: true
          description: When the rate was FETCHED upstream, not when it was used.
    CreatePayment:
      type: object
      required: [orderId]
      description: >
        Supply EXACTLY ONE pricing basis: `amount` (in crypto) or `fiatAmount` +
        `fiatCurrency`. Supplying both is a 400 — two prices would mean two
        answers to what the customer owes.
      properties:
        amount:
          type: string
          example: "50.00"
          description: Amount in `asset`. Omit when pricing in fiat.
        fiatAmount:
          type: string
          example: "5000"
          description: >
            Price in `fiatCurrency`. Converted to `asset` once, at creation, and
            frozen onto the payment (see FiatQuote). Returns 503 if no rate can
            be obtained at all — retry, or price directly in crypto.
        fiatCurrency:
          type: string
          example: INR
          description: >
            ISO-4217 code. Must be one this gateway can quote — see GET /rates.
            Required with `fiatAmount`.
        orderId:     { type: string, example: order_789 }
        description: { type: string, example: "Order #789 — Pro plan" }
        network:
          type: string
          enum: [BEP20, TRC20, ERC20, BTC]
          default: BEP20
          description: >
            Chain to settle on. Omit for BEP20 (back-compat). TRC20 requires the
            gateway to have Tron enabled, else the request returns 400.
    Invoice:
      type: object
      description: >
        A fiat-denominated document with line items. It mints no public token of
        its own: the customer-facing capability is `token`, which belongs to the
        invoice's single-use payment link.
      properties:
        id:       { type: string, example: inv_01HZY... }
        number:   { type: string, example: INV-0001, description: Sequential WITHIN this merchant }
        customerName:  { type: string, nullable: true }
        customerEmail: { type: string, nullable: true }
        currency: { type: string, example: INR }
        subtotal: { type: string, example: "3000.00" }
        taxAmount: { type: string, example: "450.00" }
        total:    { type: string, example: "3450.00" }
        notes:    { type: string, nullable: true }
        dueDate:  { type: string, format: date, nullable: true }
        status:   { type: string, enum: [draft, open, paid, void] }
        overdue:
          type: boolean
          description: >
            DERIVED from `dueDate` at read time, never stored — no job has to age
            a row into it, so an invoice cannot get stuck waiting for a clock.
        asset:    { type: string, nullable: true, description: Pinned settlement asset, or null to let the customer choose }
        network:  { type: string, nullable: true }
        token:
          type: string
          nullable: true
          description: >
            The pay link's token; the URL is {PUBLIC_PANEL_URL}/pay/{token}.
            Null while the invoice is a draft.
        paymentId: { type: string, nullable: true, description: The payment that settled it }
        subscriptionId: { type: string, nullable: true }
        cycleNumber:    { type: integer, nullable: true }
        issuedAt: { type: string, format: date-time, nullable: true }
        sentAt:   { type: string, format: date-time, nullable: true }
        paidAt:   { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        items:
          type: array
          description: Empty in LIST responses; populated by GET /invoices/{id}.
          items: { $ref: '#/components/schemas/InvoiceItem' }
    InvoiceItem:
      type: object
      properties:
        id:          { type: string }
        description: { type: string }
        quantity:    { type: string, example: "2.5", description: May be fractional }
        unitPrice:   { type: string, example: "200" }
        amount:      { type: string, example: "500.00", description: quantity × unitPrice, stored }
    CreateInvoice:
      type: object
      required: [currency, items]
      properties:
        currency: { type: string, example: INR, description: Fiat code — see GET /rates }
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: object
            required: [description, unitPrice]
            properties:
              description: { type: string }
              quantity:    { type: string, default: "1" }
              unitPrice:   { type: string }
        customerName:  { type: string }
        customerEmail: { type: string, format: email }
        notes:    { type: string }
        dueDate:  { type: string, format: date, example: "2026-08-22" }
        taxAmount: { type: string }
        asset:    { type: string, description: Pin a settlement asset; requires `network` }
        network:  { type: string, enum: [BEP20, TRC20, ERC20, BTC] }
        issue:
          type: boolean
          default: true
          description: Mint the pay link immediately. False saves a draft instead.
    Subscription:
      type: object
      description: >
        A recurring billing SCHEDULE. Each cycle issues an invoice the customer
        chooses to pay — there is no stored mandate to pull funds from.
      properties:
        id:    { type: string, format: uuid }
        title: { type: string, example: Pro plan }
        description:   { type: string, nullable: true }
        customerName:  { type: string, nullable: true }
        customerEmail: { type: string, nullable: true }
        currency: { type: string, example: INR }
        amount:   { type: string, example: "1499" }
        asset:    { type: string, nullable: true }
        network:  { type: string, nullable: true }
        intervalUnit:  { type: string, enum: [day, week, month, year] }
        intervalCount: { type: integer, example: 1 }
        status:
          type: string
          enum: [active, paused, canceled, needs_attention]
          description: >
            `needs_attention` means the plan fell further behind schedule than
            `maxCyclesBehind` allows, so billing stopped rather than emitting a
            burst of back-dated invoices. Resume it to restart on a clean
            schedule; the missed cycles are not billed.
        nextRunAt:
          type: string
          format: date-time
          description: >
            Advanced by exactly one interval from its own previous value, never
            from now(), so a worker outage does not shift every future billing
            date.
        totalCycles:   { type: integer, nullable: true, description: Null = bill indefinitely }
        cyclesBilled:  { type: integer }
        dueDays:  { type: integer, description: Days from issue to due date on each invoice }
        autoSend: { type: boolean, description: Email each invoice; requires customerEmail }
        lastInvoiceId: { type: string, nullable: true }
        lastRunAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateSubscription:
      type: object
      required: [title, amount, currency, intervalUnit]
      properties:
        title:    { type: string }
        description: { type: string }
        amount:   { type: string, example: "1499" }
        currency: { type: string, example: INR, description: Fiat code; each cycle's crypto amount is quoted when it is paid }
        intervalUnit:  { type: string, enum: [day, week, month, year] }
        intervalCount: { type: integer, default: 1, minimum: 1, maximum: 366 }
        customerName:  { type: string }
        customerEmail: { type: string, format: email }
        asset:    { type: string }
        network:  { type: string, enum: [BEP20, TRC20, ERC20, BTC] }
        startAt:  { type: string, format: date-time, description: When the FIRST invoice is minted. Defaults to now. }
        totalCycles: { type: integer }
        dueDays:  { type: integer, default: 7 }
        autoSend: { type: boolean }
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Dashboard login (merchant or admin). Returns JWT (+ MFA challenge for admins).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string }
                password: { type: string }
                mfaToken: { type: string, description: TOTP, required for admins }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:  { type: string }
                  refreshToken: { type: string }
                  mfaRequired:  { type: boolean }
                  emailVerified:
                    type: boolean
                    description: >
                      false for a self-registered merchant who has not clicked
                      their confirmation link. They can sign in, but the client
                      is still `pending` and cannot transact.
                  clientStatus: { type: string, nullable: true }
...
  /payment-links:
    get:
      tags: [Payment links]
      summary: List this merchant's hosted-checkout links
      security: [{ BearerAuth: [] }, { ApiKey: [], Signature: [], Timestamp: [] }, { ApiKeyBearer: [] }]
      responses:
        "200": { description: OK }
    post:
      tags: [Payment links]
      summary: Create a shareable checkout link
      description: >
        Returns a `token`. The customer-facing URL is
        `<panel>/pay/<token>`. Omit `amount` for an open amount (tips,
        donations). Pinning `asset` also requires `network` — the same symbol on
        two chains is two different assets.
      security: [{ BearerAuth: [] }, { ApiKey: [], Signature: [], Timestamp: [] }, { ApiKeyBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:       { type: string, maxLength: 120 }
                description: { type: string, maxLength: 500 }
                amount:      { type: string, description: Omit for an open amount }
                asset:       { type: string }
                network:     { type: string, enum: [BEP20, TRC20, ERC20, BTC] }
                reusable:    { type: boolean, default: true }
                maxUses:     { type: integer, minimum: 1 }
                expiresAt:   { type: string, format: date-time }
      responses:
        "201": { description: Created }
        "400": { description: Invalid amount, or an asset this gateway cannot settle }
...
  /rates:
    get:
      tags: [Rates]
      summary: Exchange rates for every fiat currency this gateway can price in
      description: >
        Unauthenticated, like /networks and /assets: it publishes public market
        prices and this gateway's currency list, and reveals nothing about any
        merchant.

        `source` and `ageSeconds` are part of the contract, not debug output. A
        caller displaying a converted price is entitled to know it is looking at
        a stale quote — the gateway deliberately serves a slightly old rate
        rather than failing, and says so.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  rates:
                    type: object
                    description: rates[FIAT][SYMBOL] = price of one SYMBOL in FIAT.
                    example: { USD: { USDT: "0.999498" }, INR: { USDT: "95.09" } }
                  currencies: { type: array, items: { type: string }, example: [USD, INR, EUR] }
                  source:     { type: string, example: coingecko, description: "`coingecko:stale` when served past freshness" }
                  ageSeconds: { type: integer, example: 12 }
                  fetchedAt:  { type: string, format: date-time }
        "503":
          description: >
            No rate is available and the last cached one is too old to quote
            from. Payments priced directly in crypto are unaffected.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /invoices:
    get:
      tags: [Invoices]
      summary: List invoices
      security: [{ BearerAuth: [] }, { ApiKey: [], Signature: [], Timestamp: [] }, { ApiKeyBearer: [] }]
      parameters:
        - { in: query, name: status, schema: { type: string, enum: [draft, open, paid, void] } }
        - { in: query, name: limit, schema: { type: integer, default: 100, maximum: 200 } }
      responses:
        "200":
          description: OK. `items` is empty here — use GET /invoices/{id} for line items.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Invoice' } }
    post:
      tags: [Invoices]
      summary: Create an invoice
      description: >
        Invoicing is money-IN, so an API key of either mode may do it. Issuing
        mints a single-use payment link whose token is the customer's pay URL.
      security: [{ BearerAuth: [] }, { ApiKey: [], Signature: [], Timestamp: [] }, { ApiKeyBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateInvoice' }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Invoice' }
        "400": { description: Bad request }
...
  /subscriptions:
    get:
      tags: [Subscriptions]
      summary: List subscriptions
      security: [{ BearerAuth: [] }, { ApiKey: [], Signature: [], Timestamp: [] }, { ApiKeyBearer: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Subscription' } }
    post:
      tags: [Subscriptions]
      summary: Create a recurring billing plan
      security: [{ BearerAuth: [] }, { ApiKey: [], Signature: [], Timestamp: [] }, { ApiKeyBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSubscription' }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subscription' }
        "400": { description: Bad request }
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a refresh token for a new access token
      responses:
        "200": { description: OK }
        "401": { description: Refresh token invalid or expired }
...
  /payments:
    post:
      tags: [Payments]
      summary: Create a payment (generates a fresh BEP20 or TRC20 deposit address + QR)
      security: [{ ApiKey: [], Signature: [], Timestamp: [] }]
      parameters:
        - in: header
          name: Idempotency-Key
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePayment' }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Payment' }
        "400": { description: "Unsupported/disabled network", content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        "401": { description: Bad signature, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    get:
      tags: [Payments]
      summary: List payments (paginated)
      security: [{ ApiKey: [], Signature: [], Timestamp: [] }]
      parameters:
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: page,   schema: { type: integer, default: 1 } }
        - { in: query, name: limit,  schema: { type: integer, default: 25 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Payment' } }
                  page: { type: integer }
                  total: { type: integer }
...
  /payouts:
    get:
      tags: [Payouts]
      summary: List payouts
      description: >
        Readable with any credential. Note the asymmetry with POST below:
        reading settlement history is a read, but INITIATING one moves money and
        so is restricted to HMAC-signed keys.
      security: [{ BearerAuth: [] }, { ApiKey: [], Signature: [], Timestamp: [] }, { ApiKeyBearer: [] }]
      parameters:
        - { in: query, name: page,  schema: { type: integer, default: 1 } }
        - { in: query, name: limit, schema: { type: integer, default: 20 } }
      responses:
        "200": { description: OK }
    post:
      tags: [Payouts]
      summary: Request a payout to the configured payout wallet
      security: [{ ApiKey: [], Signature: [], Timestamp: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: { type: string }
                network:
                  type: string
                  enum: [BEP20, TRC20, ERC20, BTC]
                  default: BEP20
                  description: >
                    Chain to settle on. TRC20 requires a payout_wallet_trc20 to be
                    configured and Tron enabled server-side, else returns 400.
      responses:
        "202": { description: Accepted (queued) }
        "400": { description: "No wallet for network / amount exceeds balance", content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...